Splitting of JS Files
