+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 ## UNRELEASED
 - Add capability to set a custom script path for the scripts/styles by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1557
 - Added missing tailwind background colour class for when hovering over the clear button in dark mode by @slakbal in https://github.com/rappasoft/laravel-livewire-tables/pull/1553
+- Split custom JS into individual files by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1558
 
 
 ## [v3.1.3] - 2023-11-03
